+++ R/sync.R
@@ -257,9 +257,9 @@ unlock <- function(socket) invisible(.Call(rnng_socket_unlock, socket))
 #' @param aio an Aio (object of class 'sendAio', 'recvAio' or 'ncurlAio').
 #'
 #' @details This function is identical to \code{\link{call_aio}} except that it
-#'     is user-interruptible. If interrupted, the aio is stopped upon garbage
-#'     collection, and hence the aio may return an 'errorValue' 20 'Operation
-#'     canceled' if it was still unresolved at the time.
+#'     is user-interruptible. If interrupted, the aio is stopped upon the next
+#'     garbage collection event, and hence may return an 'errorValue' 20
+#'     'Operation canceled' if it remains unresolved by that time.
 #'
 #' @return The passed object (invisibly).
 #'