+++ NEWS.md
@@ -1,4 +1,4 @@
-# nanonext 0.10.4.9000 (development)
+# nanonext 0.10.4.9001 (development)
 
 #### New Features
 
@@ -9,6 +9,7 @@
 * Restricts `nextmode()` to register custom serialization and unserialization functions for external pointer objects for simplicity.
 * `until()` updated to be identical to `.until()`, returning FALSE instead of TRUE if the timeout has been reached.
 * `pipe_notify()` arguments 'add', 'remove' and 'flag' now default to FALSE instead of TRUE for easier selective specification of the events to signal.
+* Upgrades bundled 'libmbedtls' to v 3.5.1.
 
 # nanonext 0.10.4